windows merged
