+++ artemis-core/src/artifacts/mod.rs
@@ -3,5 +3,3 @@ pub(crate) mod os;
 
 #[cfg(target_family = "unix")]
 pub(crate) mod macos_collection;
-
-pub(crate) mod windows_collection;
