+++ artemis-core/src/artifacts/os/windows/accounts/users.rs
@@ -268,13 +268,13 @@ fn get_sid(data: &[u8]) -> nom::IResult<&[u8], String> {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use crate::artifacts::os::windows::accounts::users::{
         get_flags, get_sid, parse_user_data, parse_user_info, UacFlags,
     };
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_parser_user_info() {
         let test_path = "C:\\Windows\\System32\\config\\SAM";
         let results = parse_user_info(&test_path).unwrap();
