+++ artemis-core/src/artifacts/os/windows/amcache/parser.rs
@@ -179,6 +179,7 @@ fn adjust_id(id: &str, count: usize) -> String {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use crate::{
         artifacts::os::windows::{
@@ -202,7 +203,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_grab_amcache() {
         let options = AmcacheOptions { alt_file: None };
         let result = grab_amcache(&options).unwrap();
@@ -210,7 +210,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_amcache_file() {
         let result = amcache_file(&'C').unwrap();
         assert!(result.len() > 10);
