+++ artemis-core/src/artifacts/os/windows/bits/background.rs
@@ -192,6 +192,7 @@ fn parse_carve(data: &[u8], is_legacy: bool) -> WinBits {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::parse_ese_bits;
     use crate::{
@@ -209,7 +210,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_parse_legacy_bits() {
         let results = parse_legacy_bits(&'C', false).unwrap();
         assert_eq!(results.bits.is_empty(), true);
