+++ artemis-core/src/artifacts/os/windows/error.rs
@@ -7,8 +7,6 @@ pub(crate) enum WinArtifactError {
     Ntfs,
     Output,
     Serialize,
-    Process,
-    File,
     Shimdb,
     Registry,
     Format,
@@ -40,8 +38,6 @@ impl fmt::Display for WinArtifactError {
             WinArtifactError::Ntfs => write!(f, "Failed to parse NTFS"),
             WinArtifactError::Output => write!(f, "Failed to output data"),
             WinArtifactError::Serialize => write!(f, "Artemis failed serialize artifact data"),
-            WinArtifactError::Process => write!(f, "Failed to parse Procsses"),
-            WinArtifactError::File => write!(f, "Failed to get file listing"),
             WinArtifactError::Shimdb => write!(f, "Failed to parse Shimdb"),
             WinArtifactError::Registry => write!(f, "Failed to parse Registry"),
             WinArtifactError::Format => write!(f, "Bad format option"),
