+++ artemis-core/src/artifacts/os/windows/ese/pages/longvalue.rs
@@ -215,6 +215,7 @@ fn parse_long_value_child<'a>(
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::parse_long_value;
     use crate::{
@@ -227,7 +228,6 @@ mod tests {
     use std::{collections::HashMap, path::PathBuf};
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_parse_long_value() {
         let mut test_location = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
         test_location.push("tests\\test_data\\windows\\ese\\win10\\longvalue_page.raw");
