+++ artemis-core/src/artifacts/os/windows/eventlogs/parser.rs
@@ -163,6 +163,7 @@ fn read_eventlogs(path: &str, output: &mut Output, filter: &bool) -> Result<(), 
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{alt_eventlogs, default_eventlogs, grab_eventlogs, read_directory, read_eventlogs};
     use crate::{structs::artifacts::os::windows::EventLogsOptions, structs::toml::Output};
@@ -186,7 +187,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_grab_eventlogs() {
         let options = EventLogsOptions { alt_file: None };
         let mut output = output_options("eventlog_temp", "local", "./tmp", true);
@@ -196,7 +196,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_default_eventlogs() {
         let mut output = output_options("eventlog_temp", "local", "./tmp", true);
