+++ artemis-core/src/artifacts/os/windows/prefetch/parser.rs
@@ -89,6 +89,7 @@ fn read_prefetch(path: &str) -> Result<Prefetch, PrefetchError> {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{custom_prefetch_path, grab_prefetch};
     use crate::{
@@ -98,7 +99,6 @@ mod tests {
     use std::path::PathBuf;
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_grab_prefetch() {
         let options = PrefetchOptions { alt_dir: None };
         let _ = grab_prefetch(&options).unwrap();
