+++ artemis-core/src/artifacts/os/windows/recyclebin/parser.rs
@@ -95,13 +95,13 @@ pub(crate) fn grab_recycle_bin_path(path: &str) -> Result<RecycleBin, RecycleBin
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{grab_recycle_bin, grab_recycle_bin_path};
     use crate::structs::artifacts::os::windows::RecycleBinOptions;
     use std::path::PathBuf;
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_grab_recycle_bin() {
         let options = RecycleBinOptions { alt_file: None };
         let _ = grab_recycle_bin(&options).unwrap();
