+++ artemis-core/src/artifacts/os/windows/registry/helper.rs
@@ -131,6 +131,7 @@ pub(crate) fn lookup_sk_info(path: &str, sk_offset: i32) -> Result<SecurityKey, 
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{
         get_registry_keys, get_registry_keys_by_ref, parse_raw_registry, read_registry,
@@ -144,14 +145,12 @@ mod tests {
     use std::{collections::HashMap, path::PathBuf};
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_read_registry() {
         let buffer = read_registry("C:\\Windows\\appcompat\\Programs\\Amcache.hve").unwrap();
         assert!(buffer.len() > 10000)
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_parse_raw_registry() {
         let buffer = read_registry("C:\\Windows\\appcompat\\Programs\\Amcache.hve").unwrap();
         let mut params = Params {
@@ -230,7 +229,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_get_registry_keys_by_ref() {
         let user_hives = get_user_registry_files(&'C').unwrap();
         let mut ntfs_parser = setup_ntfs_parser(&'C').unwrap();
@@ -251,7 +249,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_read_registry_ref() {
         let user_hives = get_user_registry_files(&'C').unwrap();
         let mut ntfs_parser = setup_ntfs_parser(&'C').unwrap();
