+++ artemis-core/src/artifacts/os/windows/shortcuts/parser.rs
@@ -86,6 +86,7 @@ pub(crate) fn parse_lnk_data(data: &[u8]) -> Result<ShortcutInfo, LnkError> {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{grab_lnk_directory, grab_lnk_file};
     use crate::artifacts::os::windows::shortcuts::parser::parse_lnk_data;
@@ -96,7 +97,6 @@ mod tests {
     use std::path::PathBuf;
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_recent_files() {
         let users = get_user_paths().unwrap();
         for user in users {
