+++ artemis-core/src/artifacts/os/windows/wmi/namespaces.rs
@@ -184,6 +184,7 @@ fn extract_hash_info(hash: &str) -> Option<(String, u32)> {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::{extract_hash_info, extract_namespace_data, get_classes, get_instances};
     use crate::{
@@ -202,7 +203,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     #[ignore = "Takes time to run"]
     fn test_get_classes() {
         let map_data = read_file("C:\\Windows\\System32\\wbem\\Repository\\MAPPING3.MAP").unwrap();
@@ -252,7 +252,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     #[ignore = "Takes time to run"]
     fn test_get_instances() {
         let map_data = read_file("C:\\Windows\\System32\\wbem\\Repository\\MAPPING3.MAP").unwrap();
@@ -302,7 +301,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     #[ignore = "Takes time to run"]
     fn test_extract_namespace_data() {
         let map_data = read_file("C:\\Windows\\System32\\wbem\\Repository\\MAPPING3.MAP").unwrap();
