+++ artemis-core/src/artifacts/os/windows/wmi/objects.rs
@@ -187,6 +187,7 @@ pub(crate) fn parse_record(data: &[u8]) -> nom::IResult<&[u8], ClassInfo> {
 }
 
 #[cfg(test)]
+#[cfg(target_os = "windows")]
 mod tests {
     use super::parse_page;
     use crate::{
@@ -231,7 +232,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(target_os = "windows")]
     fn test_parse_objects() {
         let maps = glob_paths("C:\\Windows\\System32\\wbem\\Repository\\MAPPING*.MAP").unwrap();
         let mut seq = 0;
