+++ artemis-core/tests/test_data/macos/groups.toml
@@ -10,5 +10,5 @@ collection_id = 1
 output = "local"
 
 [[artifacts]]
-artifact_name = "groups"
-[artifacts.groups]
+artifact_name = "groups-macos"
+[artifacts.groups_macos]
