+++ artemis-core/tests/test_data/windows/users.toml
@@ -10,6 +10,6 @@ collection_id = 1
 output = "local"
 
 [[artifacts]]
-artifact_name = "users"
-[artifacts.users]
+artifact_name = "users-windows"
+[artifacts.users_windows]
 # alt_file = "C:\\Artifacts\\SAM" # Optional
