+++ cli/src/collector/mod.rs
@@ -2,5 +2,3 @@ pub(crate) mod commands;
 
 #[cfg(target_family = "unix")]
 pub(crate) mod macos;
-
-pub(crate) mod windows;
